+++ bob/learn/tensorflow/datashuffler/Memory.py
@@ -43,7 +43,7 @@ class Memory(Base):
 
     def __init__(self, data, labels,
                  input_shape,
-                 input_dtype="float64",
+                 input_dtype="float32",
                  batch_size=1,
                  seed=10,
                  data_augmentation=None,
